// app/apps/design-systems/cloudscape/components/Table/types.ts
import { TableProps as CloudscapeTableProps } from '@cloudscape-design/components/table';
import { HeaderProps } from '@cloudscape-design/components/header';
import { CollectionPreferencesProps } from '@cloudscape-design/components/collection-preferences';

export interface BaseTableProps extends CloudscapeTableProps, Pick<HeaderProps, 'actions' | 'info' | 'description'> {
  /**
   * Disable pagination
   * */
  disablePagination?: boolean;
  /**
   * Disable preference toolbar
   * */
  disableSettings?: boolean;
  /**
   * Disable search filters
   * */
  disableFilters?: boolean;
  /**
   * Disable row select
   * */
  disableRowSelect?: boolean;
  /**
   * Options for page size in collection preferences.
   */
  collectionPreferencesProps?: Partial<CollectionPreferencesProps>;
  /**
   * The default number of rows on any given page
   * */
  defaultPageSize?: number;
  /**
   * The default index of the page that should be displayed
   * */
  defaultPageIndex?: number;
  /**
   * The heading text (together with actions, info and description props to form a Table header) <br/>
   * or the Cloudscape header component.
   */
  header?: React.ReactNode;
  /**
   * The variant of the table header.
   */
  headerVariant?: HeaderProps['variant'];
}

export interface FetchDataOptions {
  pageSize: number;
  pageIndex: number;
  filterText?: string;
  sortingField?: string;
  isDescending?: boolean;
}

export interface RemoteUpdateTableProps extends BaseTableProps {
  /**
   * Handler for updating data when table options (pageSize, pageIndex, filterText) change. <br/>
   * If the handler is not provided, data is processed automatically.
   * */
  onFetchData?: (options: FetchDataOptions) => void;
}

export interface InternalTableProps {
  collectionPreferences: CollectionPreferencesProps.Preferences;
}

export type TableProps = RemoteUpdateTableProps | BaseTableProps;

export interface ColumnDefinition<T> extends CloudscapeTableProps.ColumnDefinition<T> {}

export interface SelectionChangeDetail<T> extends CloudscapeTableProps.SelectionChangeDetail<T> {}